style: rename PopularAreas
