+++ src/components/PopularAreas.tsx
